Divide View from Controller
